% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaleidoscope.R
\name{dispatch_color}
\alias{dispatch_color}
\title{Return dispatch color based on type string}
\usage{
dispatch_color(type, colors, a = 1)
}
\arguments{
\item{type}{the generation type}

\item{colors}{data table with correlating generation type with color}

\item{a}{opacity value (default=1, opaque).}
}
\description{
Return dispatch color based on type string
}

