% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaleidoscope.R
\name{draw_chord_interchange}
\alias{draw_chord_interchange}
\title{Plot a chord diagram showing net interchange}
\usage{
draw_chord_interchange(t, iso, netinterchange, scenario = "c_RT_R30P",
  link.size = 1)
}
\arguments{
\item{t}{timestep}

\item{netinterchange}{net interchange data frame with
"time", "scenario", "Source2Sink", "value"}

\item{scenario}{scenario}
}
\description{
Plot a chord diagram showing net interchange
}

