% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_comparative_ergis_bars}
\alias{draw_comparative_ergis_bars}
\title{Stacked bar plot of ERGIS ISO generation}
\usage{
draw_comparative_ergis_bars(t, x0 = 1080/1920, x1 = 1, weight = 3)
}
\arguments{
\item{t}{timestep}
}
\description{
Plot a set stacked bar graphs of dispatch of each ERGIS ISO for
             the ERGIS scenarios
}

