% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_comparative_insight_series}
\alias{draw_comparative_insight_series}
\title{Plot a series ERGIS layout for scenario comparison in the Insight Center}
\usage{
draw_comparative_insight_series(t0, tn, prefix, max_interchange = 0)
}
\arguments{
\item{t0}{starting timestep}

\item{tn}{ending timestep}
}
\description{
Plot a series of 5760x2400 frames comparing ERGIS generation,
             net interchange and dispatch for all four scenarios
}

