% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaleidoscope.R
\name{draw_density}
\alias{draw_density}
\title{Draw heatmap of a generation type}
\usage{
draw_density(t, type, generators, generation, colors, scenario = "c_RT_R30P",
  legend_color = "white", ...)
}
\arguments{
\item{t}{timestep of interest}

\item{type}{generation type string (e.g., "Wind")}

\item{generators}{generator data frame with "Generator_Name","Node_Region","Type","lat","lon"}

\item{generation}{generation time-series as transformed by load_generation}

\item{colors}{generation type color table}

\item{ramp}{color map (defaults to rev(brewer.Greys))}

\item{shape}{underlying shapefile (defaults to ISO regions)}
}
\description{
Computes a raster from generation point sources and draws
             a heatmap of that density.
}

