% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaleidoscope.R
\name{draw_edge_group}
\alias{draw_edge_group}
\title{Plot a group igraph edges}
\usage{
draw_edge_group(df, verts, layout, arrow.scaling = 2,
  edge_color = "#FFFFFF8F")
}
\arguments{
\item{df}{Data frame with a vector of "edge" strings (e.g., "FRCC SERC")
that specify the source and destination, and a vector of
"weights" that specify the size of the interchange from source
to destination.}

\item{verts}{vector of vertex labels}

\item{layout}{vector of vertex positions}
}
\description{
Plots a group of igraph edges with a given weight. Only a
             single weight can be used, a limitation of igraph. Plot
             edges of multiple weights with multiple calls
             (see: draw_interchange)
}

