% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_ergis}
\alias{draw_ergis}
\title{Plot ERGIS generation frame}
\usage{
draw_ergis(t, scenario = "c_RT_R30P", density = "None", types = c("Hydro",
  "Coal", "Gas CC", "Wind", "CT/Gas boiler", "Other", "Pumped Storage", "PV"),
  scaling = 0.002, weight = 3, ...)
}
\arguments{
\item{t}{timestep of interest}
}
\description{
Draw a 1920x1080 frame of ERGIS with generator map and
             a dispatch stack for a single scenario
}

