% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_ergis_bars}
\alias{draw_ergis_bars}
\title{Stacked bar plot of ERGIS ISO generation}
\usage{
draw_ergis_bars(t, scenario = "c_RT_R30P", weight = 3)
}
\arguments{
\item{t}{timestep}
}
\description{
Plot a stacked bar graph of dispatch of each ERGIS ISO for
             a given scenario
}

