% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_ergis_chord}
\alias{draw_ergis_chord}
\title{Plot chord view of ERGIS data}
\usage{
draw_ergis_chord(t, max_interchange = 0, scenario = "c_RT_R30P",
  weight = 3)
}
\arguments{
\item{t}{starting timestep}

\item{max_interchange}{maximum interchange to scale the chord diagram}
}
\description{
Draw a 1920x1080 frame of ERGIS with chord diagram and
             a dispatch stack for a single scenario
}

