% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_ergis_chord_series}
\alias{draw_ergis_chord_series}
\title{Plot a series of ERGIS net interchange frames}
\usage{
draw_ergis_chord_series(t0, tn, max_interchange = 0, scenario = "c_RT_R30P",
  prefix)
}
\arguments{
\item{t0}{starting timestep}

\item{tn}{ending timestep}
}
\description{
Plot a series of ergis 1920x1080 frames with a chord diagram
             and a dispatch stack bar chart for a single scenario
}

