% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_ergis_comparative}
\alias{draw_ergis_comparative}
\title{Plot comparative view of ERGIS net interchange data}
\usage{
draw_ergis_comparative(t, max_interchange = 0)
}
\arguments{
\item{t}{starting timestep}
}
\description{
Plot a 1920x1080 frame of ERGIS net interchange for all four
             scenarios in a 2x2 pattern of chord diagrams with a dispatch
             bar chart.
}

