% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_ergis_comparative_insight}
\alias{draw_ergis_comparative_insight}
\title{Plot ERGIS layout for scenario comparison in the Insight Center}
\usage{
draw_ergis_comparative_insight(t, density = "None", max_interchange = 0,
  types = c("Hydro", "Coal", "Gas CC", "Wind", "CT/Gas boiler", "Other",
  "Pumped Storage", "PV"), scaling = 0.002, weight = 3, ...)
}
\arguments{
\item{t}{timestep of interest}
}
\description{
Plot a 5760x2400 frame comparing ERGIS generation, net interchange
             and dispatch for all four scenarios
}

