% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_ergis_comparative_map}
\alias{draw_ergis_comparative_map}
\title{Plot comparative view of ERGIS generation data}
\usage{
draw_ergis_comparative_map(t, density = "None", types = c("Hydro", "Coal",
  "Gas CC", "Wind", "CT/Gas boiler", "Other", "Pumped Storage", "PV"),
  scaling = 0.002, weight = 1, ...)
}
\arguments{
\item{t}{starting timestep}
}
\description{
Plot a 1920x1080 frame of ERGIS generation for all four
             scenarios in a 2x2 pattern with a dispatch bar chart.
}

