% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_ergis_insight}
\alias{draw_ergis_insight}
\title{Plot ERGIS layout for Insight Center}
\usage{
draw_ergis_insight(t, scenario = "c_RT_R30P", density = "None",
  types = c("Hydro", "Coal", "Gas CC", "Wind", "CT/Gas boiler", "Other",
  "Pumped Storage", "PV"), max_interchange = 0, scaling = 0.002,
  weight = 6, ...)
}
\arguments{
\item{t}{timestep of interest}
}
\description{
Plot a 5760x2400 frame of ERGIS generation, net interchange
             and dispatch for a single scenario
}

