% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_ergis_insight1_series}
\alias{draw_ergis_insight1_series}
\title{Plot insight-center layout series of ERGIS data}
\usage{
draw_ergis_insight1_series(t0, tn, max_interchange = 0,
  scenario = "c_RT_R30P", prefix)
}
\arguments{
\item{t0}{starting timestep}

\item{tn}{ending timestep}
}
\description{
Plot a series of 5760x2400 frames of ERGIS generation,
             net interchange and dispatch for a single scenario
}

