% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.parallel.R, R/parallel.R
\name{draw_ergis_parallel_series}
\alias{draw_ergis_parallel_series}
\title{Parallel Plot ERGIS data}
\usage{
draw_ergis_parallel_series(t0, tn, prefix, scenario = "c_RT_R30P",
  density = "Wind", types = c("Hydro", "Coal", "Gas CC", "Wind",
  "CT/Gas boiler", "PV"), ...)

draw_ergis_parallel_series(t0, tn, prefix, scenario = "c_RT_R30P",
  density = "Wind", types = c("Hydro", "Coal", "Gas CC", "Wind",
  "CT/Gas boiler", "PV"), ...)
}
\arguments{
\item{t0}{starting timestep}

\item{tn}{ending timestep}

\item{t0}{starting timestep}

\item{tn}{ending timestep}
}
\description{
Parallel Plot ERGIS data

Parallel Plot ERGIS data
}

