% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.R
\name{draw_ergis_series}
\alias{draw_ergis_series}
\title{Plot a series of ERGIS generation frames}
\usage{
draw_ergis_series(t0, tn, prefix, scenario = "c_RT_R30P", density = "None",
  types = c("Hydro", "Coal", "Gas CC", "Wind", "CT/Gas boiler", "Other",
  "Pumped Storage", "PV"), ...)
}
\arguments{
\item{t0}{starting timestep}

\item{tn}{ending timestep}
}
\description{
Plot a series of ergis 1920x1080 frames with a generator map
             a dispatch stack bar chart for a single scenario
}

