% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaleidoscope.R
\name{draw_generators}
\alias{draw_generators}
\title{Draw indivdiual generators}
\usage{
draw_generators(t, types, generators, generation, colors,
  scenario = "c_RT_R30P", scaling = 0.002, fill = TRUE, lx = -72,
  ly = 39.9, annontation_color = "white", legend_color = "white")
}
\arguments{
\item{t}{timestep to plot}

\item{types}{vector of generation type strings to plot}

\item{generators}{generator data frame with "Generator_Name","Node_Region","Type","lat","lon"}

\item{generation}{generation time-series as transformed by load_generation}

\item{colors}{generation type color table}

\item{scaling}{bubble scaling factor (default=0.001)}

\item{fill}{fill the generator bubble (default=TRUE)}

\item{lx}{x position of legend}

\item{ly}{y position of legend}
}
\description{
Draws individual generators as bubbles sized by generation
             and colored by generation type.
}

