% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergis.parallel.R, R/parallel.R
\name{draw_insight_parallel_series}
\alias{draw_insight_parallel_series}
\title{Parallel Plot ERGIS data}
\usage{
draw_insight_parallel_series(t0, tn, prefix, max_interchange = 0)

draw_insight_parallel_series(t0, tn, prefix, max_interchange = 0)
}
\arguments{
\item{t0}{starting timestep}

\item{tn}{ending timestep}

\item{t0}{starting timestep}

\item{tn}{ending timestep}
}
\description{
Parallel Plot ERGIS data

Parallel Plot ERGIS data
}

