% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaleidoscope.R
\name{draw_interchange}
\alias{draw_interchange}
\title{Plot ISO interchange data}
\usage{
draw_interchange(t, verts, layout, netinterchange, dispatch,
  scenario = "c_RT_R30P", arrow.scaling = 2, annontation_color = "white",
  edge_color = "white")
}
\arguments{
\item{t}{timestep of interest}

\item{verts}{vector of vertex labels}

\item{layout}{vector of vertex positions}

\item{netinterchange}{net interchange data frame with
"time", "scenario", "Source2Sink", "value"}

\item{dispatch}{regional dispatch data frame with
"Type", "time", "zone", "value", "scenario"}
}
\description{
Plot ISO interchange data
}

