% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaleidoscope.R
\name{load_generation}
\alias{load_generation}
\title{Return a transformed generation data}
\usage{
load_generation(filename, scenario)
}
\arguments{
\item{filename}{the generation csv}
}
\description{
Return transformed generation data where generation csv data
             file are in the form: "name", "YYYY-MM-DD hh:mm:ss", ...
}

