% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{plot_iso}
\alias{plot_iso}
\title{Plot the PV generations for all generators in an ISO}
\usage{
plot_iso(iso, generators, ...)
}
\arguments{
\item{iso}{the ISO's name}
}
\description{
Plot the PV generations for all generators in an ISO
}

