% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{plot_line}
\alias{plot_line}
\title{Plot a single row from the generation table}
\usage{
plot_line(name, generation, ...)
}
\arguments{
\item{name}{the generator's name}
}
\description{
Plot a single row from the generation table
}

