% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaleidoscope.R
\name{stripSmallPolys}
\alias{stripSmallPolys}
\title{Strip small polygons from complex SpatialPolygonsData}
\usage{
stripSmallPolys(poly, minarea = 0)
}
\description{
Simplifies SpatialPolygonsData by removing small polygons
}

